feat: alterando os atributos do coelho através da resposta do sistema Fuzzy

+++ SurvivalShooterECS/Assets/_Complete-Game/Scripts/Player/PlayerHealth.cs
@@ -8,7 +8,7 @@ namespace CompleteProject
     public class PlayerHealth : MonoBehaviour
     {
         public int startingHealth = 100;                            // The amount of health the player starts the game with.
-        public int currentHealth;                                   // The current health the player has.
+        public float currentHealth;                                 // The current health the player has.
         public Slider healthSlider;                                 // Reference to the UI's health bar.
         public Image damageImage;                                   // Reference to an image to flash on the screen on being hurt.
         public AudioClip deathClip;                                 // The audio clip to play when the player dies.
@@ -57,7 +57,7 @@ namespace CompleteProject
         }
 
 
-        public void TakeDamage (int amount)
+        public void TakeDamage (float amount)
         {
             // Set the damaged flag so the screen will flash.
             damaged = true;
